+++ DynamicBrickingDS.h
@@ -54,7 +54,7 @@ public:
   virtual UINTVECTOR3 GetBrickOverlapSize() const;
   UINT64VECTOR3 GetEffectiveBrickSize(const BrickKey&) const;
 
-  virtual uint64_t GetBitWidth() const;
+  virtual unsigned GetBitWidth() const;
   virtual uint64_t GetComponentCount() const;
   virtual bool GetIsSigned() const;
   virtual bool GetIsFloat() const;
