+++ KitwareConverter.cpp
@@ -52,7 +52,7 @@ KitwareConverter::KitwareConverter()
 
 bool KitwareConverter::ConvertToRAW(
   const std::string& strSourceFilename, const std::string&, bool,
-  uint64_t& iHeaderSkip, uint64_t& iComponentSize, uint64_t& iComponentCount,
+  uint64_t& iHeaderSkip, unsigned& iComponentSize, uint64_t& iComponentCount,
   bool& bConvertEndianess, bool& bSigned, bool& bIsFloat,
   UINT64VECTOR3& vVolumeSize, FLOATVECTOR3& vVolumeAspect,
   std::string& strTitle, std::string& strIntermediateFile,
@@ -231,7 +231,7 @@ bool KitwareConverter::ConvertToRAW(
 }
 
 bool KitwareConverter::ConvertToNative(const std::string& strRawFilename, const std::string& strTargetFilename, uint64_t iHeaderSkip,
-                             uint64_t iComponentSize, uint64_t iComponentCount, bool bSigned, bool bFloatingPoint,
+                             unsigned iComponentSize, uint64_t iComponentCount, bool bSigned, bool bFloatingPoint,
                              UINT64VECTOR3 vVolumeSize,FLOATVECTOR3 vVolumeAspect, bool bNoUserInteraction,
                              const bool bQuantizeTo8Bit) {
