+++ MRCConverter.cpp
@@ -96,7 +96,7 @@ MRCConverter::MRCConverter()
 
 bool MRCConverter::ConvertToNative(
      const std::string&, const std::string&, 
-     uint64_t, uint64_t, uint64_t, 
+     uint64_t, unsigned, uint64_t, 
      bool, bool, UINT64VECTOR3,
      FLOATVECTOR3, bool, 
      const bool)
@@ -106,7 +106,7 @@ bool MRCConverter::ConvertToNative(
 
 bool MRCConverter::ConvertToRAW(
      const std::string& strSourceFilename, const std::string& strTempDir,
-     bool, uint64_t& iHeaderSkip, uint64_t& iComponentSize,
+     bool, uint64_t& iHeaderSkip, unsigned& iComponentSize,
      uint64_t& iComponentCount, bool& bConvertEndianess, bool& bSigned, 
      bool& bIsFloat, UINT64VECTOR3& vVolumeSize, FLOATVECTOR3& vVolumeAspect,
      std::string&,
